Extract replay boost pad normalization
