+++ js/player/src/replay-data.ts
@@ -2,20 +2,15 @@ import type {
   BallSample,
   CameraSettings,
   PlaybackFrame,
-  RawBoostPadEvent,
   RawDemolishInfo,
   RawGoalEvent,
   RawPlayerStatEvent,
-  RawPlayerId,
   RawReplayTickMark,
   PlayerSample,
   RawBallFrame,
   RawPlayerFrame,
   RawPlayerInfo,
   RawReplayFramesData,
-  ReplayBoostPad,
-  ReplayBoostPadEvent,
-  ReplayBoostPadSize,
   ReplayModel,
   ReplayPlayerTrack,
   ReplayTickMark,
@@ -24,6 +19,12 @@ import type {
   Quaternion,
 } from "./types";
 import { getReplayHitboxSpec, inferReplayHitboxKind } from "./hitboxes";
+import {
+  buildBoostPads,
+  buildBoostPadsAsync,
+  STANDARD_SOCCAR_BOOST_PAD_COUNT,
+} from "./replay-boost-pads";
+import { buildPlayerLookup, normalizeReplayTime, playerIdToString } from "./replay-data-helpers";
 
 export interface NormalizeReplayDataOptions {
   onProgress?: (progress: number, details: NormalizeReplayProgress) => void;
@@ -60,41 +61,10 @@ const DEFAULT_CAMERA_SETTINGS: CameraSettings = {
   pitch: -4,
   fov: 110,
 };
-const BOOST_PAD_SMALL_Z = 70;
-const BOOST_PAD_BIG_Z = 73;
-const BOOST_PAD_BACK_CORNER_X = 3072;
-const BOOST_PAD_BACK_CORNER_Y = 4096;
-const BOOST_PAD_BACK_LANE_X = 1792;
-const BOOST_PAD_BACK_LANE_Y = 4184;
-const BOOST_PAD_BACK_MID_X = 940;
-const BOOST_PAD_BACK_MID_Y = 3308;
-const BOOST_PAD_CENTER_BACK_Y = 2816;
-const BOOST_PAD_SIDE_WALL_X = 3584;
-const BOOST_PAD_SIDE_WALL_Y = 2484;
-const BOOST_PAD_SIDE_LANE_X = 1788;
-const BOOST_PAD_SIDE_LANE_Y = 2300;
-const BOOST_PAD_FRONT_LANE_X = 2048;
-const BOOST_PAD_FRONT_LANE_Y = 1036;
-const BOOST_PAD_CENTER_X = 1024;
-const BOOST_PAD_CENTER_MID_Y = 1024;
-const BOOST_PAD_GOAL_LINE_Y = 4240;
-const STANDARD_SOCCAR_BOOST_PAD_COUNT = 34;
+
 const NORMALIZATION_PROGRESS_REPORT_MIN_DELTA = 0.005;
 const NORMALIZATION_PROGRESS_REPORT_FRAME_INTERVAL = Number.POSITIVE_INFINITY;
 const NORMALIZATION_ASYNC_YIELD_INTERVAL_MS = 16;
-
-function playerIdToString(playerId: RawPlayerId): string {
-  const [kind, value] = Object.entries(playerId)[0] ?? ["Unknown", "unknown"];
-  if (typeof value === "string" || typeof value === "number") {
-    return `${kind}:${value}`;
-  }
-
-  if (value && typeof value === "object") {
-    return `${kind}:${JSON.stringify(value)}`;
-  }
-
-  return `${kind}:${JSON.stringify(value)}`;
-}
 
 function normalizeVector(value: Vec3): Vec3 | null {
   const magnitude = Math.hypot(value.x, value.y, value.z);
@@ -172,10 +142,6 @@ function parseBallFrame(frame: RawBallFrame): BallSample {
     angularVelocity: rigidBody.angular_velocity ?? null,
     rotation: normalizeQuaternion(rigidBody.rotation),
   };
-}
-
-function normalizeReplayTime(rawTime: number, startTime: number): number {
-  return Math.max(0, rawTime - startTime);
 }
 
 function parsePlayerFrame(frame: RawPlayerFrame): PlayerSample {
@@ -687,77 +653,6 @@ async function buildPlayerTracksAsync(
   return players;
 }
 
-function buildPlayerLookup(players: ReplayPlayerTrack[]): Map<string, ReplayPlayerTrack> {
-  return new Map(players.map((player) => [player.id, player]));
-}
-
-function pushPad(
-  pads: ReplayBoostPad[],
-  x: number,
-  y: number,
-  z: number,
-  size: ReplayBoostPadSize,
-): void {
-  pads.push({
-    index: pads.length,
-    padId: null,
-    size,
-    position: { x, y, z },
-    events: [],
-  });
-}
-
-function pushMirrorX(
-  pads: ReplayBoostPad[],
-  x: number,
-  y: number,
-  z: number,
-  size: ReplayBoostPadSize,
-): void {
-  pushPad(pads, -x, y, z, size);
-  pushPad(pads, x, y, z, size);
-}
-
-function pushMirrorY(
-  pads: ReplayBoostPad[],
-  x: number,
-  y: number,
-  z: number,
-  size: ReplayBoostPadSize,
-): void {
-  pushPad(pads, x, -y, z, size);
-  pushPad(pads, x, y, z, size);
-}
-
-function pushMirrorXY(
-  pads: ReplayBoostPad[],
-  x: number,
-  y: number,
-  z: number,
-  size: ReplayBoostPadSize,
-): void {
-  pushMirrorX(pads, x, -y, z, size);
-  pushMirrorX(pads, x, y, z, size);
-}
-
-function buildStandardSoccarBoostPads(): ReplayBoostPad[] {
-  const pads: ReplayBoostPad[] = [];
-
-  pushMirrorY(pads, 0, BOOST_PAD_GOAL_LINE_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorXY(pads, BOOST_PAD_BACK_LANE_X, BOOST_PAD_BACK_LANE_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorXY(pads, BOOST_PAD_BACK_CORNER_X, BOOST_PAD_BACK_CORNER_Y, BOOST_PAD_BIG_Z, "big");
-  pushMirrorXY(pads, BOOST_PAD_BACK_MID_X, BOOST_PAD_BACK_MID_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorY(pads, 0, BOOST_PAD_CENTER_BACK_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorXY(pads, BOOST_PAD_SIDE_WALL_X, BOOST_PAD_SIDE_WALL_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorXY(pads, BOOST_PAD_SIDE_LANE_X, BOOST_PAD_SIDE_LANE_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorXY(pads, BOOST_PAD_FRONT_LANE_X, BOOST_PAD_FRONT_LANE_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorY(pads, 0, BOOST_PAD_CENTER_MID_Y, BOOST_PAD_SMALL_Z, "small");
-  pushMirrorX(pads, BOOST_PAD_SIDE_WALL_X, 0, BOOST_PAD_BIG_Z, "big");
-  pushMirrorX(pads, BOOST_PAD_CENTER_X, 0, BOOST_PAD_SMALL_Z, "small");
-
-  return pads;
-}
-
 function buildBallFrames(
   raw: RawReplayFramesData,
   progressTracker?: NormalizeReplayProgressTracker,
@@ -798,202 +693,6 @@ async function buildBallFramesAsync(
   }
 
   return ballFrames;
-}
-
-function parseBoostPadAvailability(kind: unknown): boolean | null {
-  if (kind === "Available") {
-    return true;
-  }
-  if (kind && typeof kind === "object") {
-    if ("Available" in kind) {
-      return true;
-    }
-    if ("PickedUp" in kind) {
-      return false;
-    }
-    const taggedKind = (kind as { kind?: unknown }).kind;
-    if (taggedKind === "Available") {
-      return true;
-    }
-    if (taggedKind === "PickedUp") {
-      return false;
-    }
-  }
-  return null;
-}
-
-function parseBoostPadSize(size: unknown): ReplayBoostPadSize | null {
-  if (size === "big" || size === "Big") {
-    return "big";
-  }
-  if (size === "small" || size === "Small") {
-    return "small";
-  }
-  return null;
-}
-
-function inferBoostPadSize(events: RawBoostPadEvent[]): ReplayBoostPadSize | null {
-  let lastPickupTime: number | null = null;
-  for (const event of events) {
-    const available = parseBoostPadAvailability(event.kind);
-    if (available === false) {
-      lastPickupTime = event.time;
-      continue;
-    }
-    if (available === true && lastPickupTime !== null) {
-      return event.time - lastPickupTime >= 7 ? "big" : "small";
-    }
-  }
-
-  return null;
-}
-
-function buildBoostPads(
-  raw: RawReplayFramesData,
-  players: ReplayPlayerTrack[],
-  startTime: number,
-  progressTracker?: NormalizeReplayProgressTracker,
-): ReplayBoostPad[] {
-  const playersById = buildPlayerLookup(players);
-  const eventsByPadId = new Map<string, RawBoostPadEvent[]>();
-
-  for (const event of raw.boost_pad_events ?? []) {
-    const availability = parseBoostPadAvailability(event.kind);
-    if (availability === null) {
-      progressTracker?.advance();
-      continue;
-    }
-    const bucket = eventsByPadId.get(event.pad_id);
-    if (bucket) {
-      bucket.push(event);
-    } else {
-      eventsByPadId.set(event.pad_id, [event]);
-    }
-    progressTracker?.advance();
-  }
-
-  const rawPads = raw.boost_pads;
-  if (!rawPads || rawPads.length === 0) {
-    progressTracker?.advance(STANDARD_SOCCAR_BOOST_PAD_COUNT);
-    return buildStandardSoccarBoostPads();
-  }
-
-  const sortedPads = [...rawPads].sort((left, right) => left.index - right.index);
-  const pads = new Array<ReplayBoostPad>(sortedPads.length);
-
-  for (let index = 0; index < sortedPads.length; index += 1) {
-    const pad = sortedPads[index]!;
-    const padId = typeof pad.pad_id === "string" ? pad.pad_id : null;
-    const rawEvents = padId ? [...(eventsByPadId.get(padId) ?? [])] : [];
-    const size =
-      parseBoostPadSize(pad.size) ??
-      inferBoostPadSize(rawEvents) ??
-      (pad.position.z >= 72 ? "big" : "small");
-
-    const sortedEvents = rawEvents.sort((left, right) => left.time - right.time);
-    const events = new Array<ReplayBoostPadEvent>(sortedEvents.length);
-
-    for (let eventIndex = 0; eventIndex < sortedEvents.length; eventIndex += 1) {
-      const event = sortedEvents[eventIndex]!;
-      const playerId = event.player ? playerIdToString(event.player) : null;
-      events[eventIndex] = {
-        time: normalizeReplayTime(event.time, startTime),
-        frame: event.frame,
-        available: parseBoostPadAvailability(event.kind) ?? true,
-        playerId,
-        playerName: playerId ? (playersById.get(playerId)?.name ?? playerId) : null,
-      };
-    }
-
-    pads[index] = {
-      index: pad.index,
-      padId,
-      size,
-      position: pad.position,
-      events,
-    };
-    progressTracker?.advance();
-  }
-
-  return pads;
-}
-
-async function buildBoostPadsAsync(
-  raw: RawReplayFramesData,
-  players: ReplayPlayerTrack[],
-  startTime: number,
-  progressTracker: AsyncNormalizeReplayProgressTracker,
-): Promise<ReplayBoostPad[]> {
-  const playersById = buildPlayerLookup(players);
-  const eventsByPadId = new Map<string, RawBoostPadEvent[]>();
-
-  for (const event of raw.boost_pad_events ?? []) {
-    const availability = parseBoostPadAvailability(event.kind);
-    if (availability === null) {
-      if (progressTracker.advance()) {
-        await progressTracker.yieldToMainThread();
-      }
-      continue;
-    }
-    const bucket = eventsByPadId.get(event.pad_id);
-    if (bucket) {
-      bucket.push(event);
-    } else {
-      eventsByPadId.set(event.pad_id, [event]);
-    }
-    if (progressTracker.advance()) {
-      await progressTracker.yieldToMainThread();
-    }
-  }
-
-  const rawPads = raw.boost_pads;
-  if (!rawPads || rawPads.length === 0) {
-    if (progressTracker.advance(STANDARD_SOCCAR_BOOST_PAD_COUNT)) {
-      await progressTracker.yieldToMainThread();
-    }
-    return buildStandardSoccarBoostPads();
-  }
-
-  const sortedPads = [...rawPads].sort((left, right) => left.index - right.index);
-  const pads = new Array<ReplayBoostPad>(sortedPads.length);
-
-  for (let index = 0; index < sortedPads.length; index += 1) {
-    const pad = sortedPads[index]!;
-    const padId = typeof pad.pad_id === "string" ? pad.pad_id : null;
-    const rawEvents = padId ? [...(eventsByPadId.get(padId) ?? [])] : [];
-    const size =
-      parseBoostPadSize(pad.size) ??
-      inferBoostPadSize(rawEvents) ??
-      (pad.position.z >= 72 ? "big" : "small");
-
-    const sortedEvents = rawEvents.sort((left, right) => left.time - right.time);
-    const events = new Array<ReplayBoostPadEvent>(sortedEvents.length);
-
-    for (let eventIndex = 0; eventIndex < sortedEvents.length; eventIndex += 1) {
-      const event = sortedEvents[eventIndex]!;
-      const playerId = event.player ? playerIdToString(event.player) : null;
-      events[eventIndex] = {
-        time: normalizeReplayTime(event.time, startTime),
-        frame: event.frame,
-        available: parseBoostPadAvailability(event.kind) ?? true,
-        playerId,
-        playerName: playerId ? (playersById.get(playerId)?.name ?? playerId) : null,
-      };
-    }
-
-    pads[index] = {
-      index: pad.index,
-      padId,
-      size,
-      position: pad.position,
-      events,
-    };
-    if (progressTracker.advance()) {
-      await progressTracker.yieldToMainThread();
-    }
-  }
-
-  return pads;
 }
 
 function createTimelineEventId(prefix: string, frame: number, suffix: string): string {
